% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_wrap_publication.R
\name{label_wrap_publication}
\alias{label_wrap_publication}
\title{wrap long label for ggplot}
\usage{
label_wrap_publication(width)
}
\arguments{
\item{width}{default width for label}
}
\description{
wrap long label for ggplot
}
\examples{
scale_x_discrete(labels = label_wrap_Publication(20))
}
