% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_publication.R
\name{theme_publication}
\alias{theme_publication}
\title{ggplot theme for publication}
\usage{
theme_publication(base_size = 12)
}
\arguments{
\item{base_size}{default size of font}
}
\description{
This function allows you to adjust your plots for publication
}
\examples{
theme_Publication()
}
\keyword{cats}
