% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{ggplot theme for publication}
\usage{
theme_Publication(base_size = 12, base_family = "helvetica")
}
\arguments{
\item{base_size}{default size of font}

\item{base_family}{default font family}
}
\description{
This function allows you to adjust your plots for publication
}
\examples{
theme_Publication()
}
\keyword{cats}
